+++ index.html
@@ -16,6 +16,7 @@
             position: relative;
       "
     ></div>
+    <div id="loading" style="width: 100%; width: 95vh"></div>
     <script class="user-imports"></script>
   </body>
 </html>
