+++ src/ui/Loading.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <div>animweb is loading</div>
+</template>
+
+<style></style>
+
