+++ src/ui/Slider.vue
@@ -28,4 +28,111 @@ onMounted(() => {
   </div>
 </template>
 
-<style></style>
+<style>
+.slider {
+  position: absolute;
+  background: #fff;
+  border: solid 0.1rem #ccc;
+  border-radius: 0.5rem;
+  padding: 0.5rem 1rem;
+  max-height: 8rem;
+  height: 8rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.slider input[type='range'] {
+  -webkit-appearance: none;
+  appearance: none;
+  background: transparent;
+  cursor: pointer;
+  margin-top: 1rem;
+}
+
+.slider input[type='text'] {
+  width: 4rem;
+  margin: 0 0.5rem;
+  margin-top: 1rem;
+  border: solid 0.1rem #ccc;
+  border-radius: 0.5rem;
+  padding: 0.5rem;
+  -moz-appearance: textfield;
+  appearance: textfield;
+  text-align: center;
+}
+
+.slider input[type='number']::-webkit-outer-spin-button,
+.slider input[type='number']::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
+
+.slider-input-container {
+  display: flex;
+  align-items: center;
+}
+
+.slider .slider-title-container {
+  text-align: center;
+}
+
+.slider .slider-title {
+  font-size: 1.3rem;
+  color: rgba(0, 0, 0, 0.8);
+}
+
+.slider-separator {
+  width: 100%;
+  height: 0.1rem;
+  background: #ccc;
+  margin: 1rem 0;
+  margin-top: 2rem;
+  margin-bottom: 1rem;
+}
+
+/***** Track Styles *****/
+/***** Chrome, Safari, Opera, and Edge Chromium *****/
+.slider input::-webkit-slider-runnable-track {
+  background: #ccc;
+  height: 0.5rem;
+  border-radius: 0.1rem;
+}
+
+/******** Firefox ********/
+.slider input::-moz-range-track {
+  background: #ccc;
+  height: 0.5rem;
+  border-radius: 0.1rem;
+}
+
+/***** Thumb Styles *****/
+/***** Chrome, Safari, Opera, and Edge Chromium *****/
+.slider input::-webkit-slider-thumb {
+  -webkit-appearance: none;
+  /* Override default look */
+  appearance: none;
+  margin-top: -0.75rem;
+  /* Centers thumb on the track */
+  background-color: #fff;
+  height: 2rem;
+  width: 2rem;
+  border-radius: 1rem;
+  border: solid 0.1rem #aaa;
+}
+
+/***** Thumb Styles *****/
+/***** Firefox *****/
+.slider input::-moz-range-thumb {
+  border: none;
+  /*Removes extra border that FF applies*/
+  border-radius: 0;
+  /*Removes default border-radius that FF applies*/
+  background-color: #fff;
+  height: 2rem;
+  width: 2rem;
+  border: solid 0.1rem #aaa;
+  border-radius: 1rem;
+}
+</style>
